一些自己的修改
1.密码加密
2.databales增加每页显示条数，保存database状态
3.修复添加与更新form被验证之后再次被打开时验证状态还在
4.数据库字段如果带空格在更新时显示数据被截断，如果此时更新，则数据被修改
5.修改用户管理页面非admin用户只显示自己的用户与密码信息，密码信息为md5加salt之后的字符
5.login支持回车
6.导出表格
7.其他

+++ flask_web.py
@@ -1,7 +1,9 @@
+#coding:utf-8
 from flask import Flask,request,render_template,redirect,url_for,session
 from config import db_config,page_config
 from dbutil.dbutil import DB
-# import requests
+from dbutil.user import User
+import requests
 import json
 import sys
 reload(sys)
@@ -11,8 +13,8 @@ app.secret_key = '\xca\x0c\x86\x04\x98@\x02b\x1b7\x8c\x88]\x1b\xd7"+\xe6px@\xc3#
 db = DB(host=db_config['host'], mysql_user=db_config['user'], mysql_pass=db_config['passwd'], \
                 mysql_db=db_config['db'])
 page_config.setdefault('favicon','/static/img/favicon.ico')
-page_config.setdefault('title','Woniu-cmdb')
-page_config.setdefault('brand_name','Woniu-cmdb')
+page_config.setdefault('title','CMDB')
+page_config.setdefault('brand_name','CMDB')
 
 @app.route('/login',methods=['GET','POST'])
 def login():
@@ -21,15 +23,9 @@ def login():
     if request.method == "POST":
         name = request.form.get('username')
         passwd = request.form.get('password')
-        print name
-        print passwd
         obj = {"result":1}
         if name and passwd:
-            sql = 'select * from user where username="%s" and password="%s"'%(name,passwd)
-            print sql
-            cur = db.execute(sql)
-            # print cur.fetchone()
-            if cur.fetchone():
+            if User(name,passwd).check():
                 obj["result"] = 0
                 session['username'] = name
         return json.dumps(obj)                
@@ -48,34 +44,62 @@ def render(template):
         return redirect('/login')
 @app.route('/addapi', methods=['POST'])
 def addapi():
+    if 'username' in session:
+        if session['username'] != "guest":
+            obj = request.form.to_dict()
+            table = obj.pop('action_type')
+            if table=="user":
+                if session['username'] == 'admin':
+                    username=obj['username']
+                    passwd = obj['password']
+                    values = obj.values()
+                    if User(username,passwd).add():
+                        res = {'result':'ok'}
+                        return json.dumps(res)
+                    else:
+                        res= {'result':'添加用户出错'}
+                        return json.dumps(res)
+                else:
+                    res = {'result':u'没有权限，请联系管理员'}
+                    return json.dumps(res)
+            else:
+                keys = obj.keys()
+                values = obj.values()
+                sql = 'insert into %s (%s) values ("%s")' % (table,','.join(keys),'","'.join(values))
+                db.execute(sql)
+                res = {'result':'ok'}
+                return json.dumps(res)
+        else:
+            res = {'result':u'没有权限，请联系管理员'}
+            return json.dumps(res)
 
-    obj = request.form.to_dict()
-    table = obj.pop('action_type')
-    keys = obj.keys()
-    values = obj.values()
-
-    sql = 'insert into %s (%s) values ("%s")' % (table,','.join(keys),'","'.join(values))
-    print sql
-    db.execute(sql)
-    res = {'result':'ok'}
-    return json.dumps(res)
 @app.route('/delapi', methods=['POST'])
 def delapi():
-
-    obj = request.form.to_dict()
-    table = obj.pop('action_type')
-    table_id = obj.pop('id')
-    sql = 'delete from %s where id=%s' %(table,table_id)
-    # sql = 'insert into %s (%s) values ("%s")' % (table,','.join(keys),'","'.join(values))
-    print sql
-    db.execute(sql)
-    res = {'result':'ok'}
-    return json.dumps(res)
+    if 'username' in session:
+        if session['username'] == "admin":
+            obj = request.form.to_dict()
+            table = obj.pop('action_type')
+            table_id = obj.pop('id')
+            sql = 'delete from %s where id=%s' %(table,table_id)
+            # sql = 'insert into %s (%s) values ("%s")' % (table,','.join(keys),'","'.join(values))
+            db.execute(sql)
+            res = {'result':'ok'}
+            return json.dumps(res)
+        else:
+            res = {'result':u'没有权限，请联系管理员'}
+            return json.dumps(res)
 
 @app.route('/listapi')
 def listapi():
     action_type = request.args.get('action_type')
-    sql = 'select * from '+action_type
+    if action_type =="user":
+        username = session['username']
+        if username !="admin":
+            sql = 'select * from %s where username="%s"' % (action_type,username)
+        else:
+            sql = 'select * from '+action_type
+    else:
+        sql = 'select * from '+action_type
     cur = db.execute(sql)
     res = {"result":cur.fetchall()}
     return json.dumps(res)
@@ -88,17 +112,32 @@ def updateapi():
     arr = []
     for key,val in obj.items():
         arr.append('%s="%s"'%(key,val))
-    print arr
-    keys = obj.keys()
-    values = obj.values()
-    sql = 'update %s set '%(table)
+    if table=="user":
+        for key,val in obj.items():
+            if key=="username":
+                username=val
+            if key=="password":
+                passwd=val
+        if username and passwd:
+            User(username,passwd).update(passwd)
+            res = {'result':'ok'}
+            return json.dumps(res)
+    else:
+        username = session['username']
+        if username!="guest":
+            keys = obj.keys()
+            values = obj.values()
+            sql = 'update %s set '%(table)
+            sql += ','.join(arr)
+            sql += ' where id='+table_id
+            print sql
+            db.execute(sql)
+            res = {'result':'ok'}
+            return json.dumps(res)
+        else:
+            res = {'result':u'没有权限，请联系管理员'}
+            return json.dumps(res)
 
-    sql += ','.join(arr)
-    sql += ' where id='+table_id
-    print sql
-    db.execute(sql)
-    res = {'result':'ok'}
-    return json.dumps(res)
 
 @app.route('/')
 def index():
